Added some unit tests for ScriptModule returning something else than a single tensor.
